fixed vi close command
git-svn-id: http://svn.xp-dev.com/svn/wxextension@4347 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/syncped/src/frame.cpp
@@ -509,6 +509,13 @@ void Frame::OnCommand(wxCommandEvent& event)
   case wxID_CLOSE:
     if (editor != NULL)
     {
+      if (!AllowClose(
+        m_NotebookWithEditors->GetId(),
+        editor))
+      {
+        return;
+      }
+        
       m_NotebookWithEditors->DeletePage(editor->GetFileName().GetFullPath());
       SetTitle(wxEmptyString, wxEmptyString);
     }
